refactored test code
