+++ source/DasBlog.Tests/FunctionalTests/BrowserBasedTests/PrototypeBrowserBasedTests.cs
@@ -1,20 +1,14 @@
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Threading;
-using DasBlog.Tests.Support.Common;
-using DasBlog.SmokeTest;
-using DasBlog.Tests.Automation.Dom;
 using DasBlog.Tests.Automation.Selenium;
-using DasBlog.Tests.Automation.Selenium.Interfaces;
 using DasBlog.Tests.Support;
 using DasBlog.Tests.Support.Interfaces;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using Xunit;
 using Microsoft.Extensions.Options;
+using Xunit;
 using Xunit.Abstractions;
-using Constants = DasBlog.Tests.Support.Common.Constants;
 
 /*
  * THERE IS NO logging to the console - in the debugger the log appears in the detail of the test results
@@ -22,7 +16,7 @@ using Constants = DasBlog.Tests.Support.Common.Constants;
  * dotnet xunit cli is required to get console output
  * If I do "dotnet xunit -diagnostics" it barfs with a reference to a missing Microsoft.Extensions.Options
  */
-namespace DasBlog.Tests.FunctionalTests
+namespace DasBlog.Tests.FunctionalTests.BrowserBasedTests
 {
 	public class BrowserOptionsAccessor : IOptions<BrowserOptions>
 	{
@@ -68,6 +62,11 @@ namespace DasBlog.Tests.FunctionalTests
 			this.logger = platform.ServiceProvider.GetService<ILoggerFactory>().CreateLogger<PrototypeBrowserBasedTests>();
 		}
 
+		[Fact]
+		public void Test1()
+		{
+			Assert.True(true);
+		}
 		[Fact(Skip="")]
 		public void MinimalTest()
 		{
@@ -126,100 +125,6 @@ namespace DasBlog.Tests.FunctionalTests
 			}
 			finally
 			{
-			}
-		}
-	}
-
-	public class BrowserTestPlatform : IDisposable
-	{
-		public IWebServerRunner Runner { get; private set; }
-		public IServiceProvider ServiceProvider { get; private set; }
-		public IBrowser Browser { get; private set; }
-		public IPublisher Publisher { get; private set; }
-		public ITestExecutor TestExecutor { get; private set; }
-		public Pages Pages { get; private set; }
-		private ILogger<BrowserTestPlatform> logger;
-		private bool init;
-
-		public BrowserTestPlatform()
-		{
-			ServiceProvider = InjectDependencies();
-			Runner = ServiceProvider.GetService<IWebServerRunner>();
-			Browser = ServiceProvider.GetService<IBrowser>();
-			TestExecutor = ServiceProvider.GetService<ITestExecutor>();
-			Publisher = ServiceProvider.GetService<IPublisher>();
-			var loggerFactory = ServiceProvider
-				.GetService<ILoggerFactory>();
-			loggerFactory.AddConsole(LogLevel.Debug)
-				.AddDebug(LogLevel.Debug);
-			logger = loggerFactory.CreateLogger<BrowserTestPlatform>();
-			Pages = new Pages(Browser);
-		}
-
-		/// <summary>
-		/// completes the platform setup by activating the logger
-		/// TODO look into how the test output helper is supposed to be injected into class fixture
-		/// </summary>
-		/// <param name="testOutputHelper">did not get injected</param>
-		public void CompleteSetup(ITestOutputHelper testOutputHelper)
-		{
-			if (!init)
-			{
-				var loggerFactory = ServiceProvider.GetService<ILoggerFactory>();
-				loggerFactory.AddProvider(new XunitLoggerProvider(testOutputHelper));
-				var dasBlogInstallation = this.ServiceProvider.GetService<IDasBlogSandbox>();
-				dasBlogInstallation.Init();
-				this.Runner.RunDasBlog();
-				this.Browser.Init();
-				init = true;
-			}
-		}
-		private IServiceProvider InjectDependencies()
-		{
-			var services = new ServiceCollection();
-			services.Configure<BrowserOptions>(options =>
-			{
-				options.HomeUrl = "http://localhost:5000/";
-				options.Driver = "firefox";
-			});
-			services.Configure<DasBlogInstallationOptions>(
-				opts => opts.ContentRootPath =
-					Path.Combine(Utils.GetProjectRootDirectory(), Constants.VanillaTestData));
-			var repoPathEnvVar = Environment.GetEnvironmentVariable(Constants.DasBlogGitRepo);
-			string repoPath;
-			if (string.IsNullOrWhiteSpace(repoPathEnvVar))
-			{
-				repoPath = Utils.GetProjectRootDirectory();
-			}
-			else
-			{
-				repoPath = repoPathEnvVar;
-			}
-			services.Configure<GitVersionedFileServiceOptions>(
-				opts => opts.GitRepo = repoPath);
-			services.AddLogging();
-			services.AddSingleton<IWebServerRunner, WebServerRunner>();
-			services.AddSingleton<IBrowser, Browser>();
-			services.AddSingleton<IPublisher, Publisher>();
-			services.AddSingleton<ITestExecutor, TestExecutor>();
-			services.AddSingleton<IVersionedFileService, GitVersionedFileService>();
-			services.AddSingleton<IDasBlogSandbox, DasBlogSandbox>();
-			return services.BuildServiceProvider();
-		}
-
-		public void Dispose()
-		{
-			try
-			{
-				Runner?.Kill();
-				Browser?.Dispose();
-			}
-			catch (Exception e)
-			{
-				// cannot do any logging here.  xunit logger throws an exception complaining that there
-				// IDisposable no active test.
-				// At the same time it refuses to let the console or debug logger work.
-				throw;
 			}
 		}
 	}
