+++ source/DasBlog.Tests/FunctionalTests/TestInfrastructureTests/InfrastructureTests.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
+using DasBlog.Core.Extensions;
 using DasBlog.Tests.Support.Common;
 using DasBlog.Tests.Support;
 using DasBlog.Tests.Support.Interfaces;
@@ -24,6 +25,15 @@ namespace DasBlog.Tests.FunctionalTests.TestInfrastructureTests
 		{
 			this.platform = platform;
 			this.platform.CompleteSetup(testOutputHelper);
+		}
+
+		[Fact]
+		public void TestLogger()
+		{
+			var loggerFactory = platform.ServiceProvider.GetService<ILoggerFactory>();
+			var logger = platform.ServiceProvider.GetService<ILoggerFactory>().CreateLogger<InfrastructureTests>();
+			logger.LogInformation("TestLogger running");
+			Assert.True(true);
 		}
 		[Fact]
 		[Trait("Category", "TestInfrastructureTest")]
@@ -60,67 +70,6 @@ namespace DasBlog.Tests.FunctionalTests.TestInfrastructureTests
 //			  );
 			  ,Path.Combine(Utils.GetProjectRootDirectory(), Constants.VanillaTestData));
 			Assert.Equal(0, exitCode);
-		}
-	}
-	public class InfrastructureTestPlatform : IDisposable
-	{
-		public IServiceProvider ServiceProvider { get; private set; }
-		private ILogger<InfrastructureTestPlatform> logger;
-		private bool init;
-
-		public InfrastructureTestPlatform()
-		{
-			ServiceProvider = InjectDependencies();
-			var loggerFactory = ServiceProvider
-				.GetService<ILoggerFactory>();
-			loggerFactory.AddConsole(LogLevel.Debug)
-				.AddDebug(LogLevel.Debug);
-			logger = loggerFactory.CreateLogger<InfrastructureTestPlatform>();
-		}
-
-		/// <summary>
-		/// completes the platform setup by activating the logger
-		/// TODO look into how the test output helper is supposed to be injected into class fixture
-		/// </summary>
-		/// <param name="testOutputHelper">did not get injected into this object's constructor
-		///   so we have to shoehorn it in here</param>
-		public void CompleteSetup(ITestOutputHelper testOutputHelper)
-		{
-			if (!init)
-			{
-				var loggerFactory = ServiceProvider.GetService<ILoggerFactory>();
-				loggerFactory.AddProvider(new XunitLoggerProvider(testOutputHelper));
-				init = true;
-			}
-		}
-		private IServiceProvider InjectDependencies()
-		{
-			var services = new ServiceCollection();
-			services.Configure<ScriptRunnerOptions>(
-				opts =>
-				{
-					opts.ScriptDirectory =
-						Path.Combine(Utils.GetProjectRootDirectory(),Constants.ScriptsRelativePath);
-					opts.ScriptTimeout = Constants.DefaultScriptTimeout;
-					if (Int32.TryParse(Environment.GetEnvironmentVariable(Constants.DasBlogTestScriptTimeout),
-						out var envScriptTimeout))
-					{
-						opts.ScriptTimeout = envScriptTimeout;
-					}
-					opts.ScriptExitTimeout = Constants.DefaultScriptExitTimeout;
-					if (Int32.TryParse(Environment.GetEnvironmentVariable(Constants.DasBlogTestScriptExitTimeout),
-						out var envScriptExitTimeout))
-					{
-						opts.ScriptExitTimeout = envScriptExitTimeout;
-					}
-				});
-			services.AddLogging();
-			services.AddSingleton<IScriptRunner, ScriptRunner>();
-			return services.BuildServiceProvider();
-		}
-
-		public void Dispose()
-		{
 		}
 	}
