+++ source/DasBlog.Tests/Support/DasBlogISandboxOptions.cs
@@ -1,6 +1,6 @@
 ﻿namespace DasBlog.Tests.Support
 {
-	public class DasBlogInstallationOptions
+	public class DasBlogISandboxOptions
 	{
 		public string ContentRootPath { get; set; }
 	}
